feat: orca 508 soft delete (#1523)
* chore: implement soft delete

* chore: added docs

* chore: set default to true

* chore: fix tests

+++ admin/backend/src/resource-docs/service/resource-docs.service.ts
@@ -56,9 +56,16 @@ export class ResourceDocsService extends BaseStorageFileService {
     return result.map((i) => this.mapResponse(i, rec_resource_id));
   }
 
+  /**
+   * Delete document from datatabase and S3
+   * @param rec_resource_id - Recreation resource ID
+   * @param document_id - document ID
+   * @param soft_delete - if set to true keeps the image on S3
+   */
   async delete(
     rec_resource_id: string,
     document_id: string,
+    soft_delete: boolean = true,
   ): Promise<RecreationResourceDocDto> {
     const doc = await this.prisma.recreation_resource_document.findUnique({
       where: { rec_resource_id, doc_id: document_id },
@@ -81,8 +88,10 @@ export class ResourceDocsService extends BaseStorageFileService {
         where: { rec_resource_id, doc_id: document_id },
       });
 
-      const s3Key = `documents/${rec_resource_id}/${document_id}/${doc.file_name}.${doc.extension}`;
-      await this.deleteS3FileSafely(s3Key);
+      if (!soft_delete) {
+        const s3Key = `documents/${rec_resource_id}/${document_id}/${doc.file_name}.${doc.extension}`;
+        await this.deleteS3FileSafely(s3Key);
+      }
 
       return this.mapResponse(doc, rec_resource_id);
     } catch (error) {
